Move main game logic in GameScreen class.
