+++ star-shooter/core/src/ru/ryabtsev/game/StarShooterGame.java
@@ -2,6 +2,7 @@ package ru.ryabtsev.game;
 
 import com.badlogic.gdx.Game;
 
+import ru.ryabtsev.game.screen.GameScreen;
 import ru.ryabtsev.game.screen.MenuScreen;
 
 /**
@@ -11,6 +12,6 @@ public class StarShooterGame extends Game {
 
 	@Override
 	public void create () {
-		setScreen( new MenuScreen() );
+		setScreen( new GameScreen() );
 	}
 }
